GMT_DOS-CLIENTS_WINDLOADS FEM PROOF  * ensure the crate compiles even if FEM_REPO is missing
